sukses upload

See merge request my-archery/api!94

+++ app/BLoC/App/ArcheryClub/LeftArcheryClub.php
@@ -17,9 +17,15 @@ class LeftArcheryClub extends Retrieval
 
     protected function process($parameters)
     {
-        $club_member = ClubMember::find($parameters->get('id'));
-        if(!$club_member){
-            throw new BLoCException("data not found");
+        $user_login = Auth::guard('app-api')->user();
+        $club_member = ClubMember::where('user_id', $user_login->id)->first();
+        
+        if (!$club_member) {
+            throw new BLoCException("user not found");
+        }
+
+        if ($club_member->role == 1) {
+            throw new BLoCException("user as a owner this club");
         }
 
         $club_member->delete();
@@ -27,8 +33,6 @@ class LeftArcheryClub extends Retrieval
 
     protected function validation($parameters)
     {
-        return [
-        
-        ];
+        return [];
     }
 }
